Move Agda.Main to src/main
