+++ src/full/Agda/Compiler/Builtins.hs
@@ -1,0 +1,8 @@
+module Agda.Compiler.Builtins where
+
+import Agda.Compiler.MAlonzo.Compiler (ghcBackend)
+import Agda.Compiler.JS.Compiler (jsBackend)
+import Agda.Compiler.Backend (Backend)
+
+builtinBackends :: [Backend]
+builtinBackends = [ ghcBackend, jsBackend ]
